refcator: added capacity component for facility
